优化订单API请求结构体

+++ apis/订单-获取订单列表.go
@@ -17,8 +17,8 @@ type ReqOrderListGet struct {
 }
 
 type TimeRange struct {
-	StartTime int `json:"start_time"`
-	EndTime   int `json:"end_time"`
+	StartTime int64 `json:"start_time"`
+	EndTime   int64 `json:"end_time"`
 }
 
 var _ bodyer = ReqOrderListGet{}
